+++ aderyn_core/src/ast/impls/node/structures.rs
@@ -1,21 +1,7 @@
-use super::*;
+use crate::ast::*;
 use crate::visitor::ast_visitor::*;
 use eyre::Result;
-use serde::{Deserialize, Serialize};
 use std::fmt::Display;
-
-#[derive(Clone, Debug, Deserialize, Eq, Serialize, PartialEq, Hash)]
-#[serde(rename_all = "camelCase")]
-pub struct StructDefinition {
-    pub name: String,
-    pub name_location: Option<String>,
-    pub visibility: Visibility,
-    pub members: Vec<VariableDeclaration>,
-    pub scope: NodeID,
-    pub canonical_name: Option<String>,
-    pub src: String,
-    pub id: NodeID,
-}
 
 impl Node for StructDefinition {
     fn accept(&self, visitor: &mut impl ASTConstVisitor) -> Result<()> {
