+++ aderyn_core/src/ast/impls/node/user_defined_value_types.rs
@@ -1,20 +1,8 @@
-use super::*;
+use crate::ast::*;
 use crate::visitor::ast_visitor::ASTConstVisitor;
 use crate::visitor::ast_visitor::Node;
 use eyre::Result;
-use serde::{Deserialize, Serialize};
 use std::fmt::Display;
-
-#[derive(Clone, Debug, Deserialize, Eq, Serialize, PartialEq, Hash)]
-#[serde(rename_all = "camelCase")]
-pub struct UserDefinedValueTypeDefinition {
-    pub underlying_type: TypeName,
-    pub name: String,
-    pub name_location: Option<String>,
-    pub canonical_name: Option<String>,
-    pub src: String,
-    pub id: NodeID,
-}
 
 impl Display for UserDefinedValueTypeDefinition {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
