+++ aderyn_core/src/ast/impls/node/variables.rs
@@ -1,30 +1,12 @@
-use super::*;
+use crate::ast::*;
 use crate::visitor::ast_visitor::*;
 use eyre::Result;
-use serde::{Deserialize, Serialize};
 use std::fmt::Display;
-
-#[derive(Clone, Copy, Debug, Deserialize, Serialize, PartialEq, PartialOrd, Eq, Ord, Hash)]
-#[serde(rename_all = "lowercase")]
-pub enum Mutability {
-    Immutable,
-    Mutable,
-    Constant,
-}
 
 impl Display for Mutability {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.write_fmt(format_args!("{}", format!("{self:?}").to_lowercase()))
     }
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, Serialize, PartialEq, PartialOrd, Eq, Ord, Hash)]
-#[serde(rename_all = "lowercase")]
-pub enum StateMutability {
-    NonPayable,
-    Payable,
-    View,
-    Pure,
 }
 
 impl Display for StateMutability {
@@ -33,70 +15,16 @@ impl Display for StateMutability {
     }
 }
 
-#[derive(Clone, Copy, Debug, Deserialize, Serialize, PartialEq, PartialOrd, Eq, Ord, Hash)]
-#[serde(rename_all = "lowercase")]
-pub enum Visibility {
-    Public,
-    Private,
-    Internal,
-    External,
-}
-
 impl Display for Visibility {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.write_fmt(format_args!("{}", format!("{self:?}").to_lowercase()))
     }
 }
 
-#[derive(Clone, Copy, Debug, Deserialize, Serialize, PartialEq, PartialOrd, Eq, Ord, Hash)]
-#[serde(rename_all = "lowercase")]
-pub enum StorageLocation {
-    Default,
-    Memory,
-    Calldata,
-    Storage,
-}
-
 impl Display for StorageLocation {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.write_fmt(format_args!("{}", format!("{self:?}").to_lowercase()))
     }
-}
-
-#[derive(Clone, Debug, Deserialize, Eq, Serialize, PartialEq, Hash)]
-#[serde(rename_all = "camelCase")]
-pub struct VariableDeclaration {
-    pub base_functions: Option<Vec<NodeID>>,
-    /// Marks whether or not the variable is a constant before Solidity 0.7.x.
-    ///
-    /// After 0.7.x you must use `mutability`. For cross-version compatibility use
-    /// [`VariableDeclaration::mutability()`].
-    #[serde(default)]
-    pub constant: bool,
-    pub documentation: Option<Documentation>,
-    pub function_selector: Option<String>,
-    pub indexed: Option<bool>,
-    /// Marks the variable's mutability from Solidity 0.7.x onwards.
-    /// For cross-version compatibility use [`VariableDeclaration::mutability()`].
-    #[serde(default)]
-    mutability: Option<Mutability>,
-    pub name: String,
-    pub name_location: Option<String>,
-    pub overrides: Option<OverrideSpecifier>,
-    pub scope: NodeID,
-    /// Marks whether or not the variable is a state variable before Solidity 0.7.x.
-    ///
-    /// After 0.7.x you must use `mutability`. For cross-version compatibility use
-    /// [`VariableDeclaration::mutability()`].
-    #[serde(default)]
-    pub state_variable: bool,
-    pub storage_location: StorageLocation,
-    pub type_descriptions: TypeDescriptions,
-    pub type_name: Option<TypeName>,
-    pub value: Option<Expression>,
-    pub visibility: Visibility,
-    pub src: String,
-    pub id: NodeID,
 }
 
 impl VariableDeclaration {
